move wikipedia wrapper into its own source
